CORE-5084 ARM support
Move away from Bitnami Helm charts to enable support for ARM. Provide
minimal overrides to reinforce that this chart is only intended for
development purposes. Uses KRaft for Kafka to minimise the number of
running processes. The downside of this is that SCRAM is not supported
so we use PLAIN for authentication.