CORE-20017, ES-2066: set memory limits for Kafka and postgres #34 (#34)
Previously we set no memory limits for Kafka on on the e2e test environment. This means that Kafka pods can use as much memory as they likes, until the Kubernetes node runs out and then Linux OOM killer will start killing stuff at random. Kafka is highly memory intensive and by design puts a lot of pressure on the memory system, so this is fairly likely to happen. This is my working theory for some test failures where we see that Kafka is going down for long enough to make Corda too slow to pass our e2e tests.

The Kafka docs recommend 4GB per instance, so set that as both request and limit. Once we have limits in place we may have to tweak the values. In addition we may have to take extra steps to make sure that Kafka does stick to its limit, otherwise Kubernetes will kill it, and that can sometimes involve passing settings to the JVM.

Similarly, set a memory limit for postgres. My end goal is that every pod has an appropriate memory limit.

This is change on the corda-dev branch which I'm told is what is used for e2e tests.